requirements.txt
fastapi
uvicorn
python-multipart
numpy
pillow
python-jose
pydantic
tensorflow